+++ database/migrations/2020_03_25_171102_create_menus_table.php
@@ -14,12 +14,12 @@ class CreateMenusTable extends Migration
     public function up()
     {
         Schema::create('menus', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('restaurant_id');
+            $table->increments('id');
+            $table->unsignedInteger('restaurant_id');
             $table->string('name', 50);
             $table->string('description')->nullable();
             $table->string('image')->nullable();
-            $table->unsignedBigInteger('category_id')->nullable();
+            $table->unsignedInteger('category_id')->nullable();
             $table->unsignedDecimal('price', 8, 2);
             $table->boolean('available')->default(true);
             $table->timestamp('added_on', 0);
