+++ database/migrations/2020_04_27_001328_create_orders_table.php
@@ -14,10 +14,10 @@ class CreateOrdersTable extends Migration
     public function up()
     {
         Schema::create('orders', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('user_id')->nullable();
-            $table->unsignedBigInteger('restaurant_id')->nullable();
-            $table->unsignedBigInteger('driver_id')->nullable();
+            $table->increments('id');
+            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedInteger('restaurant_id')->nullable();
+            $table->unsignedInteger('driver_id')->nullable();
             $table->unsignedInteger('total_items_qty');
             $table->unsignedDecimal('billing_subtotal');
             $table->unsignedDecimal('billing_delivery');
